Fix 64-bit compiler warnings.

+++ Library/Source/Nano/Core/NDebug.cpp
@@ -135,7 +135,7 @@ void NDebug::LogMessage(const char *thePath, UInt32 lineNum, const NStringUTF8 &
 
 	// Construct the message
 	fileName = mShowPath ? thePath : GetFileName(thePath);
-	snprintf(thePrefix, kPrefixBufferSize, "[%lu] %s:%ld: ", NTimeUtilities::GetBootTimeMS(), fileName, lineNum);
+	snprintf(thePrefix, kPrefixBufferSize, "[%lu] %s:%ld: ", (unsigned long) NTimeUtilities::GetBootTimeMS(), fileName, (long) lineNum);
 
 	finalMsg  = thePrefix;
 	finalMsg += theMsg;
